refactor: Split AdminController to separate file
