+++ shop_data.json
@@ -3,7 +3,7 @@
     {
       "id": 1,
       "name": "Bàn 1",
-      "status": "NO_USE"
+      "status": "USED"
     },
     {
       "id": 2,
