+++ src/main/resources/com/group13/coffeemanagement/AccountTab.fxml
@@ -1,0 +1,87 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane xmlns="http://javafx.com/javafx/23" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.group13.coffeemanagement.controller.AccountTabController">
+    <children>
+        <AnchorPane prefHeight="400.0" prefWidth="920.0" style="-fx-background-color: #e5b78d;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="91.0">
+            <children>
+                <TableView fx:id="accountTableView" editable="true" layoutY="-5.0" prefHeight="400.0" prefWidth="489.0">
+                    <columns>
+                        <TableColumn fx:id="accountIdCol" prefWidth="47.19997787475586" text="ID" />
+                        <TableColumn fx:id="accountUserNameCol" prefWidth="155.20001220703125" text="Tên tài khoản" />
+                        <TableColumn fx:id="accountNameCol" maxWidth="-1.0" minWidth="200.0" prefWidth="-1.0" text="Tên Hiển Thị" />
+                        <TableColumn fx:id="accountRoleCol" maxWidth="-1.0" minWidth="80.0" prefWidth="334.4000244140625" text="Role" />
+                    </columns>
+                </TableView>
+                <VBox layoutX="582.0" layoutY="14.0" prefHeight="248.0" prefWidth="280.0">
+                    <children>
+                        <HBox alignment="CENTER" prefHeight="44.0" prefWidth="280.0" spacing="100.0">
+                            <children>
+                                <Label text="ID" />
+                                <TextField fx:id="accountIdField" editable="false" />
+                            </children>
+                        </HBox>
+                        <HBox alignment="CENTER" layoutX="10.0" layoutY="10.0" prefHeight="44.0" prefWidth="280.0" spacing="45.0">
+                            <children>
+                                <Label text="Tên tài khoản">
+                                    <HBox.margin>
+                                        <Insets />
+                                    </HBox.margin>
+                                </Label>
+                                <TextField fx:id="accountUserNameField" promptText="Tên Tài Khoản" />
+                            </children>
+                        </HBox>
+                        <HBox alignment="CENTER" layoutX="10.0" layoutY="54.0" prefHeight="44.0" prefWidth="280.0" spacing="45.0">
+                            <children>
+                                <Label text="Tên Hiển Thị" />
+                                <TextField fx:id="accountNameField" promptText="Tên Hiển Thị" />
+                            </children>
+                        </HBox>
+                        <HBox alignment="CENTER" prefHeight="44.0">
+                            <children>
+                                <Label prefHeight="18.0" prefWidth="118.0" text="Mật Khẩu">
+                                    <HBox.margin>
+                                        <Insets left="10.0" />
+                                    </HBox.margin>
+                                </Label>
+                                <TextField fx:id="passwordField" prefWidth="200.0">
+                                    <HBox.margin>
+                                        <Insets right="10.0" />
+                                    </HBox.margin>
+                                </TextField>
+                                <Button fx:id="togglePass" minWidth="60.0" mnemonicParsing="false" onAction="#handleTogglePass" prefHeight="26.0" style="-fx-background-color: #a97550;" text="Show" textFill="WHITE">
+                                    <font>
+                                        <Font name="Calibri Bold" size="14.0" />
+                                    </font>
+                                </Button>
+                            </children>
+                        </HBox>
+                        <Button fx:id="saveAccountButton" maxWidth="1.7976931348623157E308" mnemonicParsing="false" onAction="#handleSaveAccount" style="-fx-background-color: #a97550;" text="Lưu" textFill="WHITE">
+                            <font>
+                                <Font name="Calibri Bold" size="16.0" />
+                            </font>
+                        </Button>
+                    </children>
+                </VBox>
+            </children>
+        </AnchorPane>
+        <HBox alignment="CENTER" prefHeight="90.0" prefWidth="466.0" spacing="20.0">
+            <children>
+                <TextField fx:id="searchAccountField" prefHeight="26.0" prefWidth="248.0" promptText="Tìm tên" />
+                <Button mnemonicParsing="false" onAction="#handleSearchAccount" prefHeight="30.0" prefWidth="86.0" style="-fx-background-color: #e5b78d;" text="Tìm" />
+                <Button mnemonicParsing="false" onAction="#handleAddAccount" prefHeight="30.0" prefWidth="86.0" style="-fx-background-color: #e5b78d;" text="Thêm" />
+                <Button layoutX="228.0" layoutY="34.0" mnemonicParsing="false" onAction="#handleDeleteAccount" prefHeight="30.0" prefWidth="86.0" style="-fx-background-color: #e5b78d;" text="Xóa" />
+            </children>
+        </HBox>
+    </children>
+</AnchorPane>
