+++ src/main/resources/com/group13/coffeemanagement/AdminScreen.fxml
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ChoiceBox?>
+<?import javafx.scene.control.DatePicker?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.Spinner?>
+<?import javafx.scene.control.Tab?>
+<?import javafx.scene.control.TabPane?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
+    prefHeight="600.0" prefWidth="1000.0" xmlns="http://javafx.com/javafx/23"
+    xmlns:fx="http://javafx.com/fxml/1"
+    fx:controller="com.group13.coffeemanagement.controller.AdminScreenController">
+    <children>
+        <TabPane layoutX="214.0" layoutY="146.0" prefHeight="200.0" prefWidth="200.0"
+            tabClosingPolicy="UNAVAILABLE" AnchorPane.bottomAnchor="40.0"
+            AnchorPane.leftAnchor="40.0" AnchorPane.rightAnchor="40.0" AnchorPane.topAnchor="50.0">
+            <tabs>
+                <Tab fx:id="billTab" text="Doanh Thu">
+                    <!-- <content>
+                        <fx:include source="BillTab.fxml" />
+                    </content> -->
+                </Tab>
+
+                <Tab fx:id="foodTab" text="Thức ăn">
+                    <!-- <content>
+                        <fx:include source="FoodTab.fxml" />
+                    </content> -->
+                </Tab>
+
+                <Tab fx:id="categoryTab" text="Doanh Mục">
+                    <!-- <content>
+                        <fx:include source="CategoryTab.fxml" />
+                    </content> -->
+                </Tab>
+
+                <Tab fx:id="tableTab" text="Bàn">
+                    <!-- <content>
+                        <fx:include source="TableTab.fxml" />
+                    </content> -->
+                </Tab>
+
+                <Tab fx:id="accountTab" text="Tài khoản">
+                    <!-- <content>
+                        <fx:include source="AccountTab.fxml" />
+                    </content> -->
+                </Tab>
+            </tabs>
+        </TabPane>
+        <HBox layoutY="14.0" style="-fx-background-color: #906d55;" AnchorPane.leftAnchor="0.0"
+            AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+            <children>
+                <Button mnemonicParsing="false" onAction="#backtoMain" prefWidth="50.0"
+                    style="-fx-background-color: #e5b78d;" text="Back" />
+            </children>
+            <padding>
+                <Insets bottom="10.0" left="10.0" top="10.0" />
+            </padding>
+        </HBox>
+        <HBox AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+            <children>
+                <Label fx:id="messageUpdateDB" alignment="CENTER_RIGHT" prefHeight="18.0"
+                    prefWidth="220.0" textAlignment="RIGHT" textFill="RED" />
+                <Button mnemonicParsing="false" onAction="#saveDB"
+                    style="-fx-background-color: #e5b78d;" text="Cập nhật cơ sở dữ liệu" />
+            </children>
+            <padding>
+                <Insets right="10.0" top="10.0" />
+            </padding>
+        </HBox>
+    </children>
+</AnchorPane>
+
