+++ src/main/resources/com/group13/coffeemanagement/BillTab.fxml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ChoiceBox?>
+<?import javafx.scene.control.DatePicker?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.Spinner?>
+<?import javafx.scene.control.Tab?>
+<?import javafx.scene.control.TabPane?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane xmlns="http://javafx.com/javafx/23" xmlns:fx="http://javafx.com/fxml/1"
+    fx:controller="com.group13.coffeemanagement.controller.BillTabController">
+    <children>
+        <Button layoutX="443.0" layoutY="27.0" mnemonicParsing="false" onAction="#handleViewButton"
+            prefHeight="40.0" prefWidth="81.0" style="-fx-background-color: #e5b78d;" text="Xem" />
+        <VBox alignment="CENTER" layoutX="89.0" layoutY="14.0" prefHeight="80.0" prefWidth="331.0">
+            <children>
+                <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0" spacing="20.0">
+                    <children>
+                        <Label text="Ngày Bắt Đầu" />
+                        <DatePicker fx:id="orderDatePicker" promptText="2024-10-22"
+                            showWeekNumbers="true" />
+                    </children>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0" spacing="20.0">
+                    <children>
+                        <Label text="Ngày Kết Thúc" />
+                        <DatePicker fx:id="paidDatePicker" showWeekNumbers="true" />
+                    </children>
+                </HBox>
+            </children>
+        </VBox>
+        <AnchorPane layoutX="10.0" layoutY="101.0" prefHeight="400.0" prefWidth="920.0">
+            <children>
+                <TableView fx:id="billTableView">
+                    <columns>
+                        <TableColumn fx:id="tableNameCol" prefWidth="75.0" text="Tên bàn" />
+                        <TableColumn fx:id="totalPriceCol" prefWidth="75.0" text="Tổng Tiền" />
+                        <TableColumn fx:id="orderDateCol" prefWidth="94.40003967285156"
+                            text="Thời gian vào" />
+                        <TableColumn fx:id="paidDateCol" prefWidth="93.59999084472656"
+                            text="Thời gian ra" />
+                        <TableColumn fx:id="discountCol" prefWidth="93.59999084472656"
+                            text="Giảm Giá" />
+                    </columns>
+                </TableView>
+            </children>
+        </AnchorPane>
+    </children>
+</AnchorPane>
