+++ src/main/resources/com/group13/coffeemanagement/FoodTab.fxml
@@ -1,0 +1,109 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ChoiceBox?>
+<?import javafx.scene.control.DatePicker?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.Spinner?>
+<?import javafx.scene.control.Tab?>
+<?import javafx.scene.control.TabPane?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane xmlns="http://javafx.com/javafx/23" xmlns:fx="http://javafx.com/fxml/1"
+    fx:controller="com.group13.coffeemanagement.controller.FoodTabController">
+    <children>
+        <AnchorPane prefHeight="400.0" prefWidth="920.0" style="-fx-background-color: #e5b78d;"
+            AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
+            AnchorPane.topAnchor="91.0">
+            <children>
+                <TableView fx:id="foodTableView" prefHeight="400.0" prefWidth="466.0">
+                    <columns>
+                        <TableColumn fx:id="foodIdCol" prefWidth="47.19997787475586" text="ID" />
+                        <TableColumn fx:id="foodNameCol" prefWidth="198.40003967285156" text="Tên" />
+                        <TableColumn fx:id="categoryNameCol" prefWidth="131.199951171875"
+                            text="Danh Mục" />
+                        <TableColumn fx:id="priceCol" minWidth="1.60003662109375"
+                            prefWidth="89.60003662109375" text="Giá" />
+                    </columns>
+                </TableView>
+                <VBox alignment="CENTER" layoutX="582.0" prefHeight="357.0" prefWidth="280.0"
+                    spacing="10.0" AnchorPane.bottomAnchor="10.0" AnchorPane.topAnchor="10.0">
+                    <children>
+                        <HBox alignment="CENTER" prefHeight="44.0" prefWidth="280.0" spacing="100.0">
+                            <children>
+                                <Label text="ID" />
+                                <TextField fx:id="foodIdField" editable="false" />
+                            </children>
+                        </HBox>
+                        <HBox alignment="CENTER" layoutX="10.0" layoutY="10.0" prefHeight="44.0"
+                            prefWidth="280.0" spacing="100.0">
+                            <children>
+                                <Label text="Tên" />
+                                <TextField fx:id="foodNameField" promptText="Tên món" />
+                            </children>
+                        </HBox>
+                        <HBox alignment="CENTER" layoutX="10.0" layoutY="54.0" prefHeight="44.0"
+                            prefWidth="280.0" spacing="100.0">
+                            <children>
+                                <Label prefHeight="18.0" prefWidth="135.0" text="Danh Mục" />
+                                <ChoiceBox fx:id="categoryChoiceBox" prefWidth="150.0" />
+                            </children>
+                        </HBox>
+                        <HBox alignment="CENTER" layoutX="10.0" layoutY="98.0" prefHeight="44.0"
+                            prefWidth="280.0" spacing="100.0">
+                            <children>
+                                <Label text="Giá" />
+                                <Spinner fx:id="priceSpinner" editable="true" />
+                            </children>
+                        </HBox>
+                        <HBox alignment="CENTER" layoutX="10.0" layoutY="55.0" prefHeight="44.0"
+                            prefWidth="280.0" spacing="20.0">
+                            <children>
+                                <Label prefHeight="18.0" prefWidth="151.0" text="Image" />
+                                <TextField fx:id="foodImgPathField" disable="true" editable="false"
+                                    prefHeight="26.0" prefWidth="242.0" promptText="file name" />
+                            </children>
+                        </HBox>
+                        <ImageView fx:id="foodImg" fitHeight="150.0" fitWidth="300.0"
+                            pickOnBounds="true" preserveRatio="true" />
+                        <Button fx:id="saveButton" maxWidth="1.7976931348623157E308"
+                            mnemonicParsing="false" onAction="#handleSaveFood"
+                            style="-fx-background-color: #a97550;" text="Lưu" textFill="WHITE">
+                            <font>
+                                <Font name="Calibri Bold" size="16.0" />
+                            </font>
+                        </Button>
+                    </children>
+                </VBox>
+                <Label layoutX="466.0" layoutY="252.0" text="(1024x1024) Chọn ảnh -&gt;&gt; ">
+                    <font>
+                        <Font name="System Bold" size="12.0" />
+                    </font>
+                </Label>
+            </children>
+        </AnchorPane>
+        <HBox alignment="CENTER" prefHeight="90.0" prefWidth="466.0" spacing="20.0">
+            <children>
+                <TextField fx:id="searchField" prefHeight="26.0" prefWidth="248.0"
+                    promptText="Tìm món ăn" />
+                <Button fx:id="searchButton" mnemonicParsing="false" onAction="#handleSearchFood"
+                    prefHeight="30.0" prefWidth="86.0" style="-fx-background-color: #e5b78d;"
+                    text="Tìm" />
+                <Button fx:id="addButton" mnemonicParsing="false" onAction="#handleAddFood"
+                    prefHeight="30.0" prefWidth="86.0" style="-fx-background-color: #e5b78d;"
+                    text="Thêm" />
+                <Button fx:id="deleteButton" layoutX="228.0" layoutY="34.0" mnemonicParsing="false"
+                    onAction="#handleDeleteFood" prefHeight="30.0" prefWidth="86.0"
+                    style="-fx-background-color: #e5b78d;" text="Xóa" />
+            </children>
+        </HBox>
+    </children>
+</AnchorPane>
