+++ src/main/resources/com/group13/coffeemanagement/mainScreen.fxml
@@ -5,6 +5,7 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ChoiceBox?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ScrollPane?>
 <?import javafx.scene.control.Spinner?>
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
@@ -22,7 +23,7 @@
          <children>
             <VBox>
                <children>
-                  <TitledPane alignment="CENTER" collapsible="false" contentDisplay="CENTER" maxWidth="300.0" text="Danh sách Bàn" textFill="#914c0f">
+                  <TitledPane alignment="CENTER" collapsible="false" contentDisplay="CENTER" maxWidth="240.0" text="Danh sách Bàn" textFill="#914c0f">
                      <content>
                         <FlowPane fx:id="danhSachBan" hgap="10.0" vgap="10.0">
                            <padding>
@@ -44,9 +45,9 @@
                   </Label>
                </children>
             </VBox>
-            <TitledPane alignment="CENTER" animated="false" collapsible="false" contentDisplay="CENTER" maxWidth="350.0" text="Danh Sách Món" textAlignment="CENTER" textFill="#914c0f">
+            <TitledPane alignment="CENTER" animated="false" collapsible="false" contentDisplay="CENTER" prefHeight="600.0" text="Danh Sách Món" textAlignment="CENTER" textFill="#914c0f">
                <content>
-                  <VBox alignment="CENTER" spacing="10.0" style="-fx-padding: 0;">
+                  <VBox alignment="TOP_CENTER" spacing="10.0" style="-fx-padding: 0;">
                      <children>
                         <HBox alignment="BOTTOM_CENTER">
                            <children>
@@ -65,21 +66,24 @@
                               <Button fx:id="addFoodButton" mnemonicParsing="false" styleClass="btn-dark" stylesheets="@../../../styles/mainScreen.css" text="Thêm Món">
                                  <font>
                                     <Font name="Calibri Bold" size="16.0" />
-                                 </font></Button>
+                                 </font>
+                                 <HBox.margin>
+                                    <Insets bottom="6.0" />
+                                 </HBox.margin></Button>
                            </children>
                            <VBox.margin>
                               <Insets top="10.0" />
                            </VBox.margin>
                         </HBox>
-                        <FlowPane fx:id="danhSachMon" hgap="10.0" prefHeight="433.0" vgap="10.0">
-                           <VBox.margin>
-                              <Insets />
-                           </VBox.margin>
-                           <padding>
-                              <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-                           </padding>
-                        </FlowPane>
-                        <HBox prefHeight="100.0" prefWidth="200.0" />
+                        <ScrollPane minHeight="400.0" pannable="true" style="-fx-font-size: 15; -fx-padding: 3;">
+                           <content>
+                              <FlowPane fx:id="danhSachMon" hgap="10.0" prefWidth="350.0" vgap="10.0">
+                                 <padding>
+                                    <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+                                 </padding>
+                              </FlowPane>
+                           </content>
+                        </ScrollPane>
                      </children>
                   </VBox>
                </content>
