+++ users.json
@@ -2,14 +2,14 @@
   "count": 2,
   "users": [
     {
-      "name": "Nguyen Van A",
+      "name": "Vuong Quang Quy",
       "userId": 1,
       "username": "admin",
       "passwordHash": "0Ik93TcZiRU64VydSpzJ8g\u003d\u003d",
       "role": "QUANLY"
     },
     {
-      "name": "Nguyen Nhan Vien",
+      "name": "Tran Van Hiep",
       "userId": 2,
       "username": "nhanvien",
       "passwordHash": "0Ik93TcZiRU64VydSpzJ8g\u003d\u003d",
